Update coreserver/src/edu.harvard.i2b2.crc/src/server/edu/harvard/i2b2/crc/dao/setfinder/querybuilder/BuildTempTableSql.java
This modification seems like a temporary patch to fix bugs from the different syntax between oracle and postgresql. These sorts of bugs come from the backslach used in concepts. When using the same sql sentence in postgresql, the backslash is duplicated. This patch removes extra backslashs and changes them into slash in the 'like' operation.